Verify Xdebug 2.6 availability to avoid issues
Since we don't have it available yet the container fails to build and
according to the maintainers it will take some time to have it. There
is a list of issues to be fixed and they also want to solve windows
compatibility bugs.

More info: https://bugs.xdebug.org/roadmap_page.php?version_id=43