Add flexible nginx location configuration
Because nginx stops after it encounters the first location that matches,
we need a flexible mechanism to insert configuration at that point. This
will provide that by including files from a specific directory.

We first copy over all possible location configurations and a script to
enable them so that later Dockerfiles can choose to enable certain
configurations on demand.