Base PHP cli Docker image on the official one
The necessary configuration files are under the src directory in
order to keep them organized.
The needed extensions and configurations are kept.
Manual removal of certain files are necessary since the official
image has some left overs, also php-cgi is removed since we are not
allowed to compile without it also because of the official image