Introduce Nginx Docker image
** This Commit breaks the build ** since the image was embed with
fpm and nginx now it's split, to avoid too many changes the build
process is fixed in the next commit.
Splitting is done to ensure proper logging and stability, before
we had supervisord as entrypoint taking care of 2 childs processes
(fpm and nginx), now each service lives in their own container and
outputs logs to the stdout, following good practices for docker
and kubernetes deployments.
This image is based on the official nginx one and provides simple
configuration management and kept the compatibility with the way
it was working before.