Make listen (backlog) configurable
Default settings stay as their system default but the listen.backlog
and listen URI are now configurable. The listen.backlog setting only
takes effect when you change the listen URI from a Unix socket to a TCP
address + port set to listen on.

Due to the nature of Unix sockets there is no request/connection
queue available, with TCP there is but at the cost of a higher
datatransfer overhead. So when a request comes in with a Unix
socket and all workers are busy the request is rejected. With
TCP and it's backlog queue requests are placed in there untill
they can be handled.