Add a regression test for FPM upstream config
In a previous https://github.com/usabilla/php-docker-template/pull/148
it appeared that some of the files being shipped by the upstream Docker
image were not being used.
This test guarantess that the default upstream configuration is now
tested and any attempt to remove the wrong configuration files should
yield in a failed test.