Add a sleep before running prometheus exporter e2e tests
This is a hotfix to get those tests working again. There seems to be a
race condition between the service in the container starting and the
tests needing them.

The permanent solution would be to check for X
seconds until the service is up and running before doing the tests and
failing if it's not up after those X seconds.